requirements.txt
# Core dependencies for AI Web Assistant

# LLM clients
openai>=1.0.0
anthropic>=0.15.0

# CLI interface
rich>=13.0.0
prompt_toolkit>=3.0.0

# Configuration and environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Retry logic and resilience
tenacity>=8.2.0

# Async file operations
aiofiles>=23.0.0

# Existing toolkit dependencies (likely already installed)
playwright>=1.40.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdownify>=0.11.0
sqlite-utils>=3.35.0
requests>=2.31.0

# Additional utilities
asyncio>=3.4.3
dataclasses>=0.6  # For Python < 3.7 compatibility
typing-extensions>=4.8.0

# Social media automation dependencies
keyring>=24.0.0
cryptography>=41.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
pylint>=3.0.0